/Final_Project/
@ctr0001
Not gonna lie I basically got a whole crash course in LaTeX and YAML to get the formatting right and to make this look kind of clean. All of the code to run the analyses is in the .rmd file that also knits the pdf. I made it so it would not translate into the pdf.